+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.6.0] - 2023-07-27
+
+- adds prettier formatting options and option to just generate entities
+
+### Added
+
+- `prettierConfig` key to config - Use a specific prettier config file
+- `usePrettier` key to config - Use prettier to format the files
+- `entitiesOnly` key to config - Export only the generated entities
+
+### Changed
+- `prettier` peer modules
+
+### Fixed
+- `ts-toolbox` dependency
+
 ## [1.4.1] - 2021-10-08
 
 - upgrades prisma dependencies to their latest 3.x versions